Split MessagesFragment.kt adapter --> MessagesAdapter.kt
